+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # v3.0.2
+* Fixed: Elasticsearch5 throwing unsupported operation exception when adding/updating a property
 * Added: Added a has method to the Query class to allow searches on all properties of a particular data type.
 * Added: Added a has method to the Query class to allow searches for a value across multiple properties.
 * Added: Implemented support all compare operators for DateOnly field types when using Elasticsearch 5.